+++ CapstoneTetris/tetromino/tetromino.hpp
@@ -15,18 +15,18 @@ class Tetromino
     int rotation = 0;
     void Render(SDL_Renderer* renderer);
     bool CheckMoveHorizontally(int direction,
-                               int (&board)[BOARD_HEIGHT][BOARD_WIDTH]);
+                               Uint8 (&board)[BOARD_HEIGHT][BOARD_WIDTH]);
     const int (&GetShape() const)[4][4] { return shapes[rotation]; }
     void MoveLeft();
     void MoveRight();
     void MoveDown();
-    bool CheckIfLanded(int (&board)[BOARD_HEIGHT][BOARD_WIDTH]);
-    void RotateClockwise(int board[BOARD_HEIGHT][BOARD_WIDTH]);
-    void RotateCounterClockwise(int board[BOARD_HEIGHT][BOARD_WIDTH]);
+    bool CheckIfLanded(Uint8 (&board)[BOARD_HEIGHT][BOARD_WIDTH]);
+    void RotateClockwise(Uint8 board[BOARD_HEIGHT][BOARD_WIDTH]);
+    void RotateCounterClockwise(Uint8 board[BOARD_HEIGHT][BOARD_WIDTH]);
     static Tetromino RandomTetromino();
 
   protected:
-      inline bool CheckValidRotation(int board[BOARD_HEIGHT][BOARD_WIDTH]);
+      inline bool CheckValidRotation(Uint8 board[BOARD_HEIGHT][BOARD_WIDTH]);
     Tetromino()
     {
         for (int i = 0; i < 4; i++)
@@ -41,7 +41,7 @@ class Tetromino
         }
     }
 
-    void SetRotationState(int state, const int rotation[4][4])
+    void SetRotationState(int state, const Uint8 rotation[4][4])
     {
         std::copy(&rotation[0][0], &rotation[0][0] + 4 * 4,
                   &shapes[state][0][0]);
